+++ pkg/resource/job.go
@@ -63,11 +63,13 @@ func (b JobBuilder) Build(obj client.Object) error {
 
 	// we recreate spec from ground only if we do not find the container job
 	if dbContainer, err := kube.FindContainer(JobContainerName, &job.Spec.Template.Spec); err != nil {
+		backoffLimit := int32(2)
 		job.Spec = kbatch.JobSpec{
 			// This field is alpha-level and is only honored by servers that enable the TTLAfterFinished feature.
 			// see https://v1-18.docs.kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#job-v1-batch
 			TTLSecondsAfterFinished: ptr.Int32(300),
 			Template:                b.buildPodTemplate(),
+			BackoffLimit:            &backoffLimit,
 		}
 	} else {
 		//if job with the container already exists we update the image only
@@ -150,7 +152,7 @@ func (b JobBuilder) MakeContainers() []corev1.Container {
 				},
 			},
 			Command: []string{"/bin/bash"},
-			Args:    []string{"-c", fmt.Sprintf("%s; sleep 150", GetTagVersionCommand)},
+			Args:    []string{"-c", fmt.Sprintf("set -eo pipefail; %s; sleep 150", GetTagVersionCommand)},
 		},
 	}
 }